requirements.txt
matplotlib==3.8.0
numpy==1.26.1
Pillow==10.1.0
torch==2.1.0+cu121
torchvision==0.16.0
